FirePrediction/FirePrediction_dashboard/requirements.txt
Flask
pytz==2013.7
PyYAML==3.12

